basic crypto functions moved to core
