+++ chain/core/src/std.rs
@@ -1,5 +1,6 @@
 mod bech32_address;
 mod code_hash;
+pub mod crypto;
 
 pub use bech32_address::Bech32Address;
 pub use code_hash::{CODE_HASH_LEN, code_hash};
