+++ chain/vm/Cargo.toml
@@ -27,8 +27,6 @@ wasmer-experimental = ["multiversx-chain-vm-executor-wasmer-experimental"]
 num-bigint = "0.4"
 num-traits = "0.2"
 hex = "0.4"
-sha2 = "0.10"
-sha3 = "0.10"
 itertools = "0.14"
 bitflags = "2.9"
 colored = "3.0"
