+++ framework/scenario/Cargo.toml
@@ -33,7 +33,6 @@ num-traits = "0.2"
 hex = "0.4"
 hex-literal = "1.0"
 log = { version = "0.4", features = ["std"] }
-sha2 = "0.10"
 serde = "1.0"
 serde_json = "1.0"
 pathdiff = "0.2"
